requirements.txt
ecdsa == 0.18.0
Pillow == 9.3.0
pip == 22.3.1
pycryptodome == 3.16.0
setuptools == 65.5.1
six == 1.16.0
wheel == 0.38.4